+++ CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 ### 1.5.1
 * automatically creating a new hashed url for theme javascript
 * visualization throws exception if data is insufficient
+* chart.dataset() can now take a new dataset
+* renamed vis.setX to vis.x()
 
 ### 1.5.0
 * simplified visualization api